knowledge_base/requirements.txt
openai
langchain
faiss-cpu
tiktoken
grpcio
grpcio-tools
fastapi
uvicorn
httpx
bs4
python-dateutil
html2text

